Priority Window
Use a reusable Priority Window instead of a Popover for each task.
