+++ errands/widgets/shared/task_toolbar/__init__.py
@@ -4,3 +4,4 @@ from errands.widgets.shared.task_toolbar.attachments_window import (
 )
 from errands.widgets.shared.task_toolbar.datetime_window import ErrandsDateTimeWindow
 from errands.widgets.shared.task_toolbar.notes_window import ErrandsNotesWindow
+from errands.widgets.shared.task_toolbar.priority_window import PriorityWindow
